requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
eth-hash==0.1.2
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
future==0.16.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
marshmallow==3.0.0b8
passlib==1.7.1
py-solc==2.1.0
web3==4.0.0
pycryptodome==3.5.1
PyMySQL==0.8.0
PyYAML==3.12
requests==2.18.4
semantic-version==2.6.0
six==1.11.0
SQLAlchemy==1.2.6
SQLAlchemy-Utils==0.33.1
toolz==0.9.0
urllib3==1.22
Werkzeug==0.14.1
PyJWT==1.6.1
